refactor: move load_altair_bootstrap fixture helper into test_utils
The load_altair_bootstrap convenience lived in light_client_spec_tests.rs
but was imported by unrelated test modules (light_client.rs and
consensus/light_client.rs, 8 call sites total) reaching into the consensus
spec-tests. Relocate it beside SpecTestLoader in test_utils/loader.rs and
have every consumer import crate::test_utils::load_altair_bootstrap.

It stays #[cfg(test)] pub(crate): test-only (so it is not dead code under
--features test-utils) and crate-internal (no external test_utils API
change). Call sites are unchanged apart from the import path.
